+++ NEWS.md
@@ -1,3 +1,7 @@
+# mlr3extralearners 0.6.1
+
+* Feat: Added learner `LearnerClassifImbalancedRandomForestSRC` (https://github.com/mlr-org/mlr3extralearners/issues/48)
+
 # mlr3extralearners 0.6.0-9000
 
 * BREAKING CHANGE: lightgbm's early stopping mechanism now uses the task's test set.